Instantiate the service once, instead of per-request
