+++ via/services/__init__.py
@@ -1,4 +1,9 @@
 """Services for Via."""
+import json
+from json import JSONDecodeError
+from pathlib import Path
+
+from via.exceptions import ConfigurationError
 from via.services.google_drive import GoogleDriveAPI
 from via.services.secure_link import SecureLinkService, has_secure_url_token
 from via.services.via_client import ViaClientService
@@ -7,12 +12,55 @@ from via.services.via_client import ViaClientService
 def includeme(config):  # pragma: no cover
     """Add services to pyramid config."""
 
+    config.register_service(
+        create_google_api(config.registry.settings), iface=GoogleDriveAPI
+    )
+
     config.register_service_factory(
         "via.services.via_client.factory", iface=ViaClientService
     )
     config.register_service_factory(
-        "via.services.google_drive.factory", iface=GoogleDriveAPI
-    )
-    config.register_service_factory(
         "via.services.secure_link.factory", iface=SecureLinkService
     )
+
+
+def create_google_api(settings):
+    """Create from Pyramid settings."""
+
+    if not settings.get("google_drive_in_python"):
+        return GoogleDriveAPI(credentials_list=None)
+
+    return GoogleDriveAPI(
+        credentials_list=load_injected_json(settings, "google_drive_credentials.json")
+    )
+
+
+def load_injected_json(settings, file_name, required=True):
+    """Load a JSON file from the env specified `DATA_DIRECTORY`.
+
+    This data is provided to us externally (by S3 at the moment) or any other
+    mechanism which causes files to exist in the directory we expect.
+
+    :param settings: A dict of Pyramid settings
+    :param file_name: Filename to load
+    :param required: Return None instead of raising if the file is missing
+    :return: Decoded JSON data
+
+    :raises ConfigurationError: If the file is required and not found or
+        malformed
+    """
+
+    data_directory: Path = settings.get("data_directory")
+    resource = data_directory / file_name
+
+    if not resource.exists():
+        if not required:
+            return None
+
+        raise ConfigurationError(f"Expected data file '{resource}' not found")
+
+    with resource.open(encoding="utf-8") as handle:
+        try:
+            return json.load(handle)
+        except JSONDecodeError as exc:
+            raise ConfigurationError(f"Invalid data file format: '{resource}'") from exc
